+++ docs/examples/comments/post.ts
@@ -1,0 +1,14 @@
+import { Client } from '@robgietema/nick';
+
+const cli = Client.initialize({ apiPath: 'http://localhost:8080' });
+const login = await cli.login({
+  data: { login: 'admin', password: 'admin' },
+});
+
+const { data } = await cli.addComment({
+  token: login.data.token,
+  path: '/events/event-1',
+  data: {
+    text: 'New comment',
+  },
+});