+++ src/seeds/profile/profile.ts
@@ -1,6 +1,7 @@
 import type { Knex } from 'knex';
 import { fileExists } from '../../helpers/fs/fs';
 import { stripI18n } from '../../helpers/i18n/i18n';
+import { omit } from 'es-toolkit/object';
 
 import models from '../../models';
 
@@ -14,7 +15,7 @@ export const seedProfile = async (
       (await import(`${profilePath}/metadata`)).default,
     );
     await Profile.deleteById(profile.id, trx);
-    await Profile.create(profile, {}, trx);
+    await Profile.create(omit(profile, ['upgrade']), {}, trx);
     console.log('Profile imported');
   }
 };